little cleanup of common logic with lua scripts
